+++ src/grafzahl/count_modes/countable.rs
