+++ src/grafzahl/count_modes/custom_modes/char.rs
@@ -2,12 +2,12 @@ use std::fmt::{Display, Formatter};
 use std::iter::Sum;
 use std::ops::Add;
 use crate::AppState;
-use crate::grafzahl::countable::Countable;
+use crate::grafzahl::count_modes::countable::Countable;
 
 #[derive(Default, Clone)]
-pub(crate) struct WordCount(u32);
+pub(crate) struct CharCount(u32);
 
-impl Add for WordCount {
+impl Add for CharCount {
     type Output = Self;
 
     fn add(self, rhs: Self) -> Self::Output {
@@ -15,40 +15,38 @@ impl Add for WordCount {
     }
 }
 
-impl Sum for WordCount {
+impl Sum for CharCount {
     fn sum<I: Iterator<Item=Self>>(iter: I) -> Self {
         iter.fold(Self::default(), |acc, num| acc + num)
     }
 }
 
-impl Display for WordCount {
+impl Display for CharCount {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.0)
     }
 }
 
-impl Countable for WordCount {
+impl Countable for CharCount {
     fn count(content: Vec<String>, _: &str, _: &AppState) -> Result<Self, String> {
         let mut count: u32 = 0;
         for line in content {
-            count += line.trim().split_whitespace().count() as u32;
+            count += line.len() as u32;
         }
         return Ok(Self(count));
     }
 
     fn display_summary(self, project_name: String) {
-        println!("Project: {} => {} Words", project_name, self);
+        println!("Project: {} => {} Character", project_name, self);
     }
 
     fn display_legend() {
-        println!("Legend: => Amount of Words");
+        println!("Legend: => Amount of Chars")
     }
 
     fn display_description() {
-        println!("Counting Mode: Word Count");
+        println!("Counting Mode: Character Count");
         println!("--------------------------------------------------");
-        println!("Count all Word within a File!");
-        println!("Something is considered a Word when it is seperated by Unicode whitespace.");
-        println!("Example: \"x = 4 + 5\" would count as 5 Words.");
+        println!("Count all Characters within a File!");
     }
 }
