+++ src/grafzahl/count_modes/custom_modes/line.rs
@@ -1,5 +1,5 @@
 use crate::AppState;
-use crate::grafzahl::countable::Countable;
+use crate::grafzahl::count_modes::countable::Countable;
 
 pub(crate) type LineCount = u32;
