+++ src/grafzahl/count_modes/description_printer.rs
@@ -1,4 +1,4 @@
-use crate::grafzahl::countable::Countable;
+use crate::grafzahl::count_modes::countable::Countable;
 
 pub(crate) fn print_description<CountMode: Countable>() {
     CountMode::display_description();
