+++ src/grafzahl/mod.rs
@@ -1,8 +1,5 @@
 pub mod languages;
 pub mod ignore_checker;
-pub mod count_mode;
-pub mod description_printer;
+pub mod count_modes;
 mod generic_counter;
 mod io_reader;
-mod countable;
-mod count_mode_lib;
